+++ src/components/context/OntologyContext.tsx
@@ -1,5 +1,5 @@
 import { createContext } from 'react';
 
-import { Ontology } from '../../../types/ontologies.model';
+import { Ontology } from '../../types/ontologies.model';
 
 export const OntologyContext = createContext<Ontology | undefined>(undefined);
