+++ src/components/context/ShapeContext.tsx
@@ -1,5 +1,5 @@
 import { createContext } from 'react';
 
-import { Shape } from '../../../types/shapes.model';
+import { Shape } from '../../types/shapes.model';
 
 export const ShapeContext = createContext<Shape | undefined>(undefined);
