+++ src/components/detailsPage/pages/ontologies/components/actions/OntologyActions.tsx
@@ -5,7 +5,7 @@ import { useNavigate } from 'react-router-dom';
 import GaiaXButton from '../../../../../../common/components/buttons/GaiaXButton';
 import Title from '../../../../../../common/components/fields/title/Title';
 import { downloadTurtleFile } from '../../../../../../services/schemaService.utils';
-import { OntologyContext } from '../../../../../index/context/OntologyContext';
+import { OntologyContext } from '../../../../../context/OntologyContext';
 
 import styles from './OntologyActions.module.css';
