+++ src/components/detailsPage/pages/shapes/ShapesDetailsPage.tsx
@@ -1,4 +1,4 @@
-import { ShapeContext } from 'components/index/context/ShapeContext';
+import { ShapeContext } from 'components/context/ShapeContext';
 import React, { FC, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useLocation } from 'react-router-dom';
