+++ src/components/detailsPage/pages/shapes/components/actions/ShapeActions.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import GaiaXButton from '../../../../../../common/components/buttons/GaiaXButton';
 import Title from '../../../../../../common/components/fields/title/Title';
 import { downloadJsonFile, downloadTurtleFile } from '../../../../../../services/schemaService.utils';
-import { ShapeContext } from '../../../../../index/context/ShapeContext';
+import { ShapeContext } from '../../../../../context/ShapeContext';
 
 import styles from './ShapeActions.module.css';
