+++ src/components/main/components/navbar/components/MenuIcon.tsx
@@ -1,7 +1,7 @@
 import Icon from '@mui/icons-material/Menu';
 import React, { FC, useContext } from 'react'
 
-import { AuthContext } from '../../../../index/context/AuthContextProvider';
+import { AuthContext } from '../../../../context/AuthContextProvider';
 
 import styles from './MenuIcon.module.css';
