+++ src/pages/details/DetailsPage.tsx
@@ -2,7 +2,7 @@
 import DetailsCard from 'components/cards/DetailsCard';
 import MapCard from 'components/cards/MapCard';
 import SidebarCard from 'components/cards/SidebarCard';
-import { AuthContext } from 'components/index/context/AuthContextProvider';
+import { AuthContext } from 'components/context/AuthContextProvider';
 import React, { useContext, useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { CypherQueryApiService as cypherQuery } from 'services/cypherQueryApiService';
